Solving leaks v2

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: dicarval <dicarval@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 14:30:24 by dicarval          #+#    #+#             */
-/*   Updated: 2024/05/14 10:36:45 by dicarval         ###   ########.fr       */
+/*   Updated: 2024/09/12 14:02:08 by dicarval         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	long int		res;
 	long int		sign;
